+++ regulatory_reporting/api_views/__init__.py
@@ -1,10 +1,11 @@
 from .flaw import FlawSRPReportMilestoneViewSet, FlawSRPReportViewSet
-from .srp_milestone import SRPReportMilestoneViewSet
+from .srp_milestone import SRPMilestoneListViewSet, SRPReportMilestoneViewSet
 from .srp_report import SRPReportViewSet
 from .upstream_notifications import UpstreamNotificationView
 
 __all__ = [
     "SRPReportViewSet",
+    "SRPMilestoneListViewSet",
     "SRPReportMilestoneViewSet",
     "FlawSRPReportViewSet",
     "FlawSRPReportMilestoneViewSet",