+++ docs/CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Upstream Maintainer Email Template (OSIDB-5086)
 - Add Upstream Maintainer Notification Payload Preparation Service (OSIDB-5087)
 - Add SRP report endpoints (OSIDB-5073)
+- Add top-level SRP milestone list endpoint with filters (OSIDB-5239)
 
 ### Changed
 - align existing workflows with workflow framework concepts